refactoring : complete
