+++ src/components/DDay/DDayBox.tsx
@@ -1,22 +1,18 @@
 import { useDragControls } from "framer-motion";
 import { useEffect, useState } from "react";
-import { useRecoilState } from "recoil";
-import { dDayListState, UserInput } from "../../recoil/dDay";
 import JuiceFont from "../common/JuiceFont";
-import SetBtn from "./SetBtn";
-import Tooltip from "./Tooltip";
+import SetBtn from "../common/SetBtn";
+import Tooltip from "../common/Tooltip";
 
 interface IDDAYS {
   dDayName: string;
   date: string;
   isSetBtn?: boolean;
-  isFirst?: boolean;
-  isDelete?: string;
+  isFirst?: number;
+  isDelete?: () => void;
 }
 
-const Dday = ({ dDayName, date, isSetBtn, isFirst, isDelete }: IDDAYS) => {
-  const [list, setList] = useRecoilState<UserInput[]>(dDayListState);
-
+const DDayBox = ({ dDayName, date, isSetBtn, isFirst, isDelete }: IDDAYS) => {
   const dragControls = useDragControls();
 
   // 오늘을 표기하기 위한 today
@@ -40,15 +36,12 @@ const Dday = ({ dDayName, date, isSetBtn, isFirst, isDelete }: IDDAYS) => {
       +lastWeek.slice(0, lastWeekDays).includes(0) +
       +lastWeek.slice(0, lastWeekDays).includes(6));
 
-  const deleteItem = (name: string | undefined) => {
-    list.length !== 1 && setList(list.filter((item) => item.dDayName !== name));
-  };
   return (
     <>
       <div className="flex justify-between w-full bg-gradient-to-r from-yellow-200 via-green-200 to-green-500 rounded-md shadow-md p-2 ">
         <div className="flex flex-col">
           <JuiceFont isBold others="mb-4">
-            {isFirst && `⭐️ `}
+            {isFirst === 0 && `⭐️ `}
 
             {dDayName}
           </JuiceFont>
@@ -67,7 +60,7 @@ const Dday = ({ dDayName, date, isSetBtn, isFirst, isDelete }: IDDAYS) => {
             <SetBtn dragControls={dragControls} />
             <button
               className="hover:bg-gray-100/[0.5] hover:rounded-lg duration-500"
-              onClick={() => deleteItem(isDelete)}
+              onClick={isDelete}
             >
               ❌
             </button>
@@ -78,4 +71,4 @@ const Dday = ({ dDayName, date, isSetBtn, isFirst, isDelete }: IDDAYS) => {
   );
 };
 
-export default Dday;
+export default DDayBox;
