+++ src/components/common/Divider.tsx
@@ -1,0 +1,5 @@
+const Divider = () => {
+  return <hr className="border-gray-300" />;
+};
+
+export default Divider;
